precarga combobox de clases y validaciones de miembros

+++ Miembros.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Data;
 using System.Data.SqlClient;
+using System.Text.RegularExpressions;
 
 namespace GYMSTATS
 {
@@ -22,6 +23,7 @@ namespace GYMSTATS
         private void Miembros_Load(object sender, EventArgs e)
         {
             dgvMiembros.DataSource = llenar_grid();
+            //
         }
 
         public DataTable llenar_grid()
@@ -39,23 +41,40 @@ namespace GYMSTATS
 
         private void btnAgregar_Click(object sender, EventArgs e)
         {
-            Conexion.Conectar();
-            string insertar = "INSERT INTO Miembros (cod_mie,tipo_nro_doc,nombre_mie,apellido_mie,fecha_nac,tel,estado,correo)VALUES(@cod_mie,@tipo_nro_doc,@nombre_mie,@apellido_mie,@fecha_nac,@tel,@estado,@correo)";
-            SqlCommand cmd1 = new SqlCommand(insertar, Conexion.Conectar());
-            cmd1.Parameters.AddWithValue("@cod_mie", txtCod.Text);
-            cmd1.Parameters.AddWithValue("@tipo_nro_doc", txtDoc.Text);
-            cmd1.Parameters.AddWithValue("@nombre_mie", txtNombre.Text);
-            cmd1.Parameters.AddWithValue("@apellido_mie", txtApellido.Text);
-            cmd1.Parameters.AddWithValue("@fecha_nac", txtNac.Text);
-            cmd1.Parameters.AddWithValue("@tel", txtTel.Text);
-            cmd1.Parameters.AddWithValue("@estado", txtEstado.Text);
-            cmd1.Parameters.AddWithValue("@correo", txtCorreo.Text);
+            //aca
 
-            cmd1.ExecuteNonQuery();
+            if (ValidarNumeroEntero(txtCod) &&
+                ValidarLetrasSinEspacios(txtNombre) &&
+                ValidarLetrasSinEspacios(txtApellido) &&
+                ValidarNumeroEntero(txtTel) &&
+                ValidarFecha(txtNac) &&
+                ValidarEmail(txtCorreo) &&
+                ValidarNoVacio(txtEstado))
+            {
+                MessageBox.Show("Validaciones son exitosas");
 
-            MessageBox.Show("Miembro agregado con exito");
+                Conexion.Conectar();
+                string insertar = "INSERT INTO Miembros (cod_mie,tipo_nro_doc,nombre_mie,apellido_mie,fecha_nac,tel,estado,correo)VALUES(@cod_mie,@tipo_nro_doc,@nombre_mie,@apellido_mie,@fecha_nac,@tel,@estado,@correo)";
+                SqlCommand cmd1 = new SqlCommand(insertar, Conexion.Conectar());
+                cmd1.Parameters.AddWithValue("@cod_mie", txtCod.Text);
+                cmd1.Parameters.AddWithValue("@tipo_nro_doc", txtDoc.Text);
+                cmd1.Parameters.AddWithValue("@nombre_mie", txtNombre.Text);
+                cmd1.Parameters.AddWithValue("@apellido_mie", txtApellido.Text);
+                cmd1.Parameters.AddWithValue("@fecha_nac", txtNac.Text);
+                cmd1.Parameters.AddWithValue("@tel", txtTel.Text);
+                cmd1.Parameters.AddWithValue("@estado", txtEstado.Text);
+                cmd1.Parameters.AddWithValue("@correo", txtCorreo.Text);
 
-            dgvMiembros.DataSource = llenar_grid();
+                cmd1.ExecuteNonQuery();
+
+                MessageBox.Show("Miembro agregado con exito");
+
+                dgvMiembros.DataSource = llenar_grid();
+            }
+            else
+            {
+                MessageBox.Show("Datos incorrectos");
+            }
         }
 
         private void dgvMiembros_CellContentClick(object sender, DataGridViewCellEventArgs e)
@@ -76,35 +95,58 @@ namespace GYMSTATS
 
         private void btnModificar_Click(object sender, EventArgs e)
         {
-            Conexion.Conectar();
-            string actualizar = "UPDATE Miembros SET tipo_nro_doc=@tipo_nro_doc, nombre_mie=@nombre_mie, apellido_mie=@apellido_mie, fecha_nac=@fecha_nac, tel=@tel, estado=@estado, correo=@correo WHERE cod_mie=@cod_mie";
-            SqlCommand cmd2 = new SqlCommand(actualizar, Conexion.Conectar());
-            cmd2.Parameters.AddWithValue("@cod_mie", txtCod.Text);
-            cmd2.Parameters.AddWithValue("@tipo_nro_doc", txtDoc.Text);
-            cmd2.Parameters.AddWithValue("@nombre_mie", txtNombre.Text);
-            cmd2.Parameters.AddWithValue("@apellido_mie", txtApellido.Text);
-            cmd2.Parameters.AddWithValue("@fecha_nac", txtNac.Text);
-            cmd2.Parameters.AddWithValue("@tel", txtTel.Text);
-            cmd2.Parameters.AddWithValue("@estado", txtEstado.Text);
-            cmd2.Parameters.AddWithValue("@correo", txtCorreo.Text);
+            if (ValidarNumeroEntero(txtCod) &&
+                ValidarLetrasSinEspacios(txtNombre) &&
+                ValidarLetrasSinEspacios(txtApellido) &&
+                ValidarNumeroEntero(txtTel) &&
+                ValidarFecha(txtNac) &&
+                ValidarEmail(txtCorreo) &&
+                ValidarNoVacio(txtEstado))
+            {
+                MessageBox.Show("Validaciones son exitosas");
 
-            cmd2.ExecuteNonQuery();
+                Conexion.Conectar();
+                string actualizar = "UPDATE Miembros SET tipo_nro_doc=@tipo_nro_doc, nombre_mie=@nombre_mie, apellido_mie=@apellido_mie, fecha_nac=@fecha_nac, tel=@tel, estado=@estado, correo=@correo WHERE cod_mie=@cod_mie";
+                SqlCommand cmd2 = new SqlCommand(actualizar, Conexion.Conectar());
+                cmd2.Parameters.AddWithValue("@cod_mie", txtCod.Text);
+                cmd2.Parameters.AddWithValue("@tipo_nro_doc", txtDoc.Text);
+                cmd2.Parameters.AddWithValue("@nombre_mie", txtNombre.Text);
+                cmd2.Parameters.AddWithValue("@apellido_mie", txtApellido.Text);
+                cmd2.Parameters.AddWithValue("@fecha_nac", txtNac.Text);
+                cmd2.Parameters.AddWithValue("@tel", txtTel.Text);
+                cmd2.Parameters.AddWithValue("@estado", txtEstado.Text);
+                cmd2.Parameters.AddWithValue("@correo", txtCorreo.Text);
 
-            MessageBox.Show("Datos actualizados con exito");
-            dgvMiembros.DataSource = llenar_grid();
+                cmd2.ExecuteNonQuery();
+
+                MessageBox.Show("Datos actualizados con exito");
+                dgvMiembros.DataSource = llenar_grid();
+            }
+            else
+            {
+                MessageBox.Show("Datos incorrectos");
+            }
         }
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
-            Conexion.Conectar();
-            string eliminar = "DELETE FROM Miembros WHERE cod_mie = @cod_mie";
-            SqlCommand cmd3 = new SqlCommand(eliminar, Conexion.Conectar());
-            cmd3.Parameters.AddWithValue("@cod_mie", txtCod.Text);
+            if (ValidarNumeroEntero(txtCod))
+            {
+                Conexion.Conectar();
+                string eliminar = "DELETE FROM Miembros WHERE cod_mie = @cod_mie";
+                SqlCommand cmd3 = new SqlCommand(eliminar, Conexion.Conectar());
+                cmd3.Parameters.AddWithValue("@cod_mie", txtCod.Text);
 
-            cmd3.ExecuteNonQuery();
+                cmd3.ExecuteNonQuery();
 
-            MessageBox.Show("Miembro eliminado con exito");
-            dgvMiembros.DataSource = llenar_grid();
+                MessageBox.Show("Miembro eliminado con exito");
+                dgvMiembros.DataSource = llenar_grid();
+            }
+            else
+            {
+                MessageBox.Show("Debe seleccionar un miembro para ser eliminado");
+            }
+            
         }
 
         private void btnLimpiar_Click(object sender, EventArgs e)
@@ -123,5 +165,38 @@ namespace GYMSTATS
         {
             this.Close();
         }
+
+        #region VALIDACIONES
+        private bool ValidarNumeroEntero(TextBox textBox)
+        {
+            int numero;
+            return int.TryParse(textBox.Text, out numero);
+        }
+
+        private bool ValidarLetrasSinEspacios(TextBox textBox)
+        {
+            return !string.IsNullOrEmpty(textBox.Text) && Regex.IsMatch(textBox.Text, "^[a-zA-Z]+$");
+        }
+
+        private bool ValidarFecha(TextBox textBox)
+        {
+            DateTime fecha;
+            if (DateTime.TryParse(textBox.Text, out fecha))
+            {
+                return fecha.Hour == 0 && fecha.Minute == 0;
+            }
+            return false;
+        }
+
+        private bool ValidarEmail(TextBox textBox)
+        {
+            return !string.IsNullOrEmpty(textBox.Text) && Regex.IsMatch(textBox.Text, @"^\w+@\w+\.\w+$");
+        }
+
+        private bool ValidarNoVacio(TextBox textBox)
+        {
+            return !string.IsNullOrWhiteSpace(textBox.Text);
+        }
+        #endregion
     }
 }
